Move ViewModels to a separate namespace and folder
