+++ InventoryManagement-BT/ViewModels/TakeInventorySearchResultsViewModel.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Web;
 
-namespace InventoryManagement_BT.Models
+namespace InventoryManagement_BT.ViewModels
 {
     public class TakeInventorySearchResultsViewModel
     {
